api/requirements.txt
asyncpg==0.30.0
fastapi==0.115.12
jose==1.0.0
pydantic==2.11.2
python-dotenv==1.1.0
python_bcrypt==0.3.2
python_jose==3.4.0
uvicorn==0.34.0
